feat(doodle): 优化 DoodleImageView 和 ClickArea 相关代码，修复涂抹时单击崩溃的问题，增加取消涂抹时删除 cache 图片的逻辑，补充注释

+++ app/src/main/java/net/taikula/autohelper/tools/FileUtils.kt
@@ -43,8 +43,14 @@ object FileUtils {
         return file
     }
 
+    /**
+     * 删除文件
+     */
     fun delete(filePath: String): Boolean {
         val file = File(filePath)
-        return file.delete()
+        if (file.exists())
+            return file.delete()
+
+        return true
     }
 }
